pyproject.toml
[project]
name = "ticket-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.14",
    "fastapi-cli>=0.0.7",
    "google-auth>=2.40.3",
    "jose>=1.0.0",
    "jupyterlab>=4.4.4",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
